refactor(parser): split expr/precedence_meta_ops.rs into cohesive submodules (§7-8)
Split the 1341-line parser/expr/precedence_meta_ops.rs (over the 500-line
limit) into a facade plus four themed submodules under
src/parser/expr/precedence_meta_ops/. All submodules are under 500 lines.
precedence_meta_ops.rs becomes a facade: mod declarations and pub(super)
re-exports preserving every external caller unchanged.

- precedence_meta_ops.rs (22)   — facade: mod decls + re-exports
- meta_bracket.rs (406)         — block_newline_terminates, parse_meta_op,
                                  BracketInfix, parse_bracket_infix_op,
                                  op_str_to_token_kind, parse_infix_func_op,
                                  strip_sequence_op + private helpers + tests
- set_ops.rs (214)              — parse_set_op, parse_negated_set_op,
                                  structural_expr
- hyper_concat.rs (308)         — parse_hyper_op, parse_hyper_func_op,
                                  hyper_op_prec, parse_hyper_rhs,
                                  concat_expr, replication_expr
- arith.rs (436)                — try_custom_infix_at_level, OpPrecedence,
                                  classify_base_op, try_bracket_op_at_level,
                                  additive_expr, multiplicative_expr,
                                  prefix_expr_with_ws_dot,
                                  autoincrement_expr,
                                  power_expr, power_expr_tight,
                                  power_expr_inner

Pure relocation. Submodule functions pub(crate); facade re-exports at
pub(super) preserving original external visibility. External callers
(precedence.rs, postfix/loop_.rs) use unchanged absolute/relative import
paths. Body refs to super::/super::super:: rewritten to absolute
crate::parser:: paths. The fn multiset is byte-identical. No behavior change.
